refactoring: update package name
