+++ jnosql-mapping/jnosql-mapping-column/src/main/java/org/eclipse/jnosql/mapping/column/AbstractMapperQuery.java
@@ -18,7 +18,7 @@ import org.eclipse.jnosql.communication.column.Column;
 import org.eclipse.jnosql.communication.column.ColumnCondition;
 import org.eclipse.jnosql.mapping.core.Converters;
 import org.eclipse.jnosql.mapping.metadata.EntityMetadata;
-import org.eclipse.jnosql.mapping.util.ConverterUtil;
+import org.eclipse.jnosql.mapping.core.util.ConverterUtil;
 
 import java.util.List;
 import java.util.stream.StreamSupport;
