+++ jnosql-mapping/jnosql-mapping-column/src/main/java/org/eclipse/jnosql/mapping/column/query/RepositoryColumnBean.java
@@ -22,7 +22,7 @@ import org.eclipse.jnosql.mapping.DatabaseType;
 import org.eclipse.jnosql.mapping.column.JNoSQLColumnTemplate;
 import org.eclipse.jnosql.mapping.metadata.EntitiesMetadata;
 import org.eclipse.jnosql.mapping.spi.AbstractBean;
-import org.eclipse.jnosql.mapping.util.AnnotationLiteralUtil;
+import org.eclipse.jnosql.mapping.core.util.AnnotationLiteralUtil;
 
 import java.lang.annotation.Annotation;
 import java.lang.reflect.Proxy;
