+++ jnosql-mapping/jnosql-mapping-core/src/test/java/org/eclipse/jnosql/mapping/core/DatabaseQualifierTest.java
@@ -14,6 +14,7 @@
  */
 package org.eclipse.jnosql.mapping.core;
 
+import org.eclipse.jnosql.mapping.DatabaseQualifier;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
