+++ jnosql-mapping/jnosql-mapping-key-value/src/main/java/org/eclipse/jnosql/mapping/keyvalue/query/AbstractKeyValueRepositoryProxy.java
@@ -19,8 +19,8 @@ import jakarta.enterprise.inject.spi.CDI;
 import jakarta.nosql.keyvalue.KeyValueTemplate;
 import org.eclipse.jnosql.mapping.DynamicQueryException;
 import org.eclipse.jnosql.mapping.query.RepositoryType;
-import org.eclipse.jnosql.mapping.repository.DynamicQueryMethodReturn;
-import org.eclipse.jnosql.mapping.repository.ThrowingSupplier;
+import org.eclipse.jnosql.mapping.core.repository.DynamicQueryMethodReturn;
+import org.eclipse.jnosql.mapping.core.repository.ThrowingSupplier;
 
 import java.lang.reflect.InvocationHandler;
 import java.lang.reflect.InvocationTargetException;
